backend/requirements.txt
fastapi
uvicorn[standard]
launchdarkly-server-sdk
apify-client
google-generativeai
Pillow
requests
python-dotenv